+++ js/Auslese.tsx
@@ -20,6 +20,7 @@ export interface AusleseProps
     type?: AusleseTypes.SelectionType;
     selections?: WeakMap<AusleseTypes.Choice, boolean>;
     onChange?: (selection: AusleseTypes.SelectedChoice[]) => void;
+    dropdownHolder?: HTMLElement;
     class?: string|null;
     placeholder?: string;
     emptyResultsMessage?: string;
@@ -53,7 +54,7 @@ export class Auslese extends Component<AusleseProps, AusleseState>
     public constructor (props: Readonly<AusleseProps>)
     {
         super(props);
-        this.dropdownHolder = document.body;
+        this.dropdownHolder = props.dropdownHolder || document.body;
         this.state = this.initState(props);
         this.onBodyClickBound = event => this.onBodyClick(event);
     }